+++ ui/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- NativeSQL translator: Added support for the percentage step
 - NativeSQL translator: Added support for the concatenate step
 
 ## [0.106.0] - 2023-07-31